requirements.txt
fastapi[all] 
torch
transformers
pillow
uvicorn
numpy 
opencv-python 
scikit-learn 
scikit-image 
matplotlib
mahotas